requirements.txt
gdown==3.13.0
numpy==1.19.2
pandas==1.2.4
pretrainedmodels==0.7.4
pytorch-lightning==1.3.8
pytorch-metric-learning==0.9.99
torch==1.8.1
torchvision==0.10.0
tqdm==4.51.0
stackprinter==0.2.5
faiss-cpu==1.7.0
umap-learn==0.5.1
matplotlib==3.3.4
scipy==1.6.0
scikit-learn==0.24.1
mlflow==1.19.0
Pillow==8.2.0